requirements.txt
absl-py==0.11.0
aiohttp==3.7.3
aiohttp-cors==0.7.0
aioredis==1.3.1
astor==0.8.1
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
blessings==1.7
cached-property==1.5.2
cachetools==4.1.1
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
colorama==0.4.4
colored==1.4.2
colorful==0.5.4
colour==0.1.5
contextvars==2.4
Cython==0.29.21
fasteners==0.15
filelock==3.0.12
future==0.18.2
gast==0.4.0
GitPython==3.0.2
glfw==2.0.0
google==3.0.0
google-api-core==1.23.0
google-auth==1.23.0
google-pasta==0.2.0
googleapis-common-protos==1.52.0
gpustat==0.6.0
GPUtil==1.4.0
grpcio==1.34.0
gym==0.12.4
h5py==3.1.0
hiredis==1.1.0
idna==2.10
idna-ssl==1.1.0
imageio==2.9.0
immutables==0.14
importlib-metadata==3.1.1
jsonschema==3.2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.3
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
monotonic==1.5
msgpack==1.0.0
mujoco-py==2.0.2.9
multidict==5.1.0
numpy==1.16.5
nvidia-ml-py3==7.352.0
olefile==0.46
opencensus==0.7.11
opencensus-context==0.1.2
prometheus-client==0.9.0
protobuf==3.14.0
psutil==5.7.3
py-spy==0.3.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyglet==1.5.11
pyrsistent==0.17.3
pytz==2020.4
PyYAML==5.3.1
ray==1.0.1
redis==3.4.1
requests==2.25.0
rsa==4.6
scipy==1.5.4
smmap==3.0.4
smmap2==3.0.1
soupsieve==2.0.1
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
torch==1.1.0
torchvision==0.3.0
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.6.3
zipp==3.4.0
gitdb2<=2.0.5

